review y funcionalidad de registrar
