+++ src/components/ToolbarComponent.vue
@@ -2,21 +2,14 @@
 import { useRoute } from 'vue-router';
 import SidebarMenu from './SidebarMenu.vue';
 import SidebarRight from './SidebarRight.vue';
-import useActuacion from '../composables/useActuacion';
-import MyCalendar from './elementos/MyCalendar.vue';
 import { ref, watch, onActivated } from 'vue';
 
 const route = useRoute();
-const { fechaCreacion, setFechaCreacion } = useActuacion();
-const showCalendar = ref(false);
+const isNewActuacion = ref(false);
 const today = ref<Date | null>(null);
 
 onActivated(() => {
   checkRoute();
-});
-
-watch(fechaCreacion, (newValue) => {
-  today.value = newValue ? new Date(newValue) : null;
 });
 
 watch(route, () => {
@@ -24,11 +17,7 @@ watch(route, () => {
 });
 
 const checkRoute = () => {
-  showCalendar.value =
-    route.name !== 'actuaciones' && route.name !== 'editActuacion';
-};
-const actualizarFechaCreacion = (fechaSeleccionada: Date) => {
-  setFechaCreacion(fechaSeleccionada);
+  isNewActuacion.value = route.name === 'newActuacion';
 };
 </script>
 
@@ -39,19 +28,18 @@ const actualizarFechaCreacion = (fechaSeleccionada: Date) => {
     </template>
 
     <template #center>
-      <div v-if="showCalendar && today" class="flex align-items-center gap-2">
-        <div class="text-3xl">Nueva Actuación</div>
-        <!-- <MyCalendar
-          v-model="today"
-          :fechaDesde="null"
-          :fechaHasta="'today'"
-          @update:modelValue="actualizarFechaCreacion"
-        /> -->
+      <div
+        v-if="route.name === 'actuaciones'"
+        class="flex align-items-center gap-2"
+      >
+        <div class="text-3xl">Actuaciones</div>
       </div>
-      <div v-else-if="showCalendar" class="p-d-flex p-flex-column text-3xl">
-        Actuaciones
+      <div v-else-if="isNewActuacion" class="p-d-flex p-flex-column text-3xl">
+        Nueva Actuación
       </div>
-      <div v-else class="p-d-flex p-flex-column text-3xl">Actuaciones</div>
+      <div v-else class="p-d-flex p-flex-column text-3xl">
+        Edición de actuación
+      </div>
     </template>
 
     <template #end>
